requirements.txt
flask==2.0.1
gunicorn
openai==0.27.2
pandas==1.3.0
numpy==1.21.0
joblib
tensorflow==2.5.0
scikeras
python-dotenv==1.0.0
scikit-learn == 1.4.2 

